feat: create routes and error page
