+++ src/main.jsx
@@ -1,14 +1,20 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import Root from "./routes/root";
-import ErrorPage from "./error-page";
+
+import Root from "./routes/index/root";
+import School from "./routes/school/school";
+import ErrorPage from "./routes/error/error-page";
 
 const router = createBrowserRouter([
   {
     path: "/",
     element: <Root />,
     errorElement: <ErrorPage />,
+  },
+  {
+    path: "/escola/:schoolId",
+    element: <School />,
   },
 ]);
