+++ src/routes/error/error-page.css
@@ -1,0 +1,41 @@
+.container__erro-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-height: 63.5vh;
+  padding-top: 4rem;
+
+  font-family: Inter, monospace;
+  color: #1c3152;
+}
+
+.container__erro-page > h1 {
+  font-size: 4rem;
+  font-weight: 600;
+  line-height: 6rem;
+  margin-bottom: 4rem;
+}
+
+.container > .container__error {
+  font-size: 1.5rem;
+  margin-bottom: 2rem;
+}
+
+ .container__button {
+  height: 3rem;
+  width: 6rem;
+  color: #fff;
+  background-color: #c1c1c1;
+  border: none;
+  font-size: 1rem;
+  font-family: Inter, monospace;
+  cursor: pointer;
+  border-radius: 10px;
+
+  margin-top: 5rem;
+}
+
+.container__button:hover{
+  background-color: #1c3152;
+  color: #fff;
+}
