+++ src/routes/school/school.jsx
@@ -1,0 +1,12 @@
+import { Header } from "../../components/header/header";
+
+function School() {
+  return (
+    <>
+      <Header />
+      <div>Schools page</div>
+    </>
+  );
+}
+
+export default School;
